requirements.txt
streamlit
pandas
openai
python-dotenv
reportlab
matplotlib
plotly